tests are picking up NUnit addin

+++ default.ps1
@@ -5,7 +5,7 @@ properties {
     $solution = "$base_dir\NJasmine.sln"
     $msbuild_Configuration = "Debug"
 	$deploySource = "$base_dir\NJasmine\bin\$msbuild_Configuration\"
-	$testDeployTarget = "$base_dir\packages\NUnit.2.5.7.10213\Tools\bin\"
+	$testDeployTarget = "$base_dir\packages\NUnit.2.5.7.10213\Tools\addins\"
 	$testDll = "$base_dir\NJasmine.Tests\bin\$msbuild_Configuration\NJasmine.Tests.dll"
 }
 
@@ -17,9 +17,11 @@ task Build {
 
 task Test -depends Build {
 
-	rm $testDeployTarget -recurse
+	if (test-path $testDeployTarget) {
+		rm $testDeployTarget -recurse
+	}
 	mkdir $testDeployTarget
-	cp $deploySource $testDeployTarget\addins -recurse
+	cp $deploySource\* $testDeployTarget -recurse
 
 	exec { .\packages\NUnit.2.5.7.10213\Tools\nunit-console.exe $testDll }
 }
